Add tests for `AuthServiceImpl`, `RoleServiceImpl`, and `UserServiceImpl`.
